Renamed step3_handler into preview_ascii
