+++ addie/menu/preview_ascii/step3_gui_handler.py
@@ -3,7 +3,7 @@ from qtpy.QtWidgets import (QFileDialog)
 import os
 
 from addie.utilities.file_handler import FileHandler
-from addie.menu.step3_handler.preview_ascii_window import PreviewAsciiWindow
+from addie.menu.preview_ascii.preview_ascii_window import PreviewAsciiWindow
 
 
 class Step3GuiHandler(object):
